chore: lint baseline to zero — 26 ruff + 36 mypy errors burned down (vibe-quant-4aiin)
Typed genome deserialization helpers, reconciliation open-event tuples,
FitnessResult/GenerationResult tuple annotations, int_param in plugin
lookbacks, isinstance guards over blind float()/int() casts, engine Any
annotations, LatencyPreset-typed CLI accessor. ruff check + mypy now clean
across 159 files.

+++ vibe_quant/overfitting/bootstrap_sharpe.py
@@ -66,7 +66,7 @@ def _sharpe_from_returns(returns: np.ndarray) -> float:
     std = float(np.std(returns, ddof=1))
     if std < 1e-10:
         return 0.0
-    return mean / std * np.sqrt(n)
+    return float(mean / std * np.sqrt(n))
 
 
 def bootstrap_sharpe_ci(
